requirements.txt
fastapi 
uvicorn
earthengine-api>=0.1.367
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
numpy
pandas
pydantic
torch 
torchvision 
shapely 
ultralytics 
Pillow
python-multipart
windows-curses